refactor: LineChart 코드 분리
- createChart
- filterChartValues
- setPeriodValues_List

- makeDebounceHandler 추가
