+++ src/utils/makeDebounceHandler.ts
@@ -1,0 +1,8 @@
+export default function makeDebouncedHandler(eventHandler: () => any, delay: number) {
+	let timer = null;
+
+	return () => {
+		if (timer) clearTimeout(timer);
+		setTimeout(eventHandler, delay);
+	};
+}
